controller/user & routes/user: refactor controller to barrel pattern & organise by subdomain
