+++ server/controllers/user.controller/index.ts
@@ -1,0 +1,4 @@
+export * from './signup';
+export * from './apiKey';
+export * from './userPreferences';
+export * from './authManagement';
