+++ server/routes/user.routes.ts
@@ -4,15 +4,28 @@ import isAuthenticated from '../utils/isAuthenticated';
 
 const router = Router();
 
+/**
+ * ===============
+ * SIGN UP
+ * ===============
+ */
 // POST /signup
 router.post('/signup', UserController.createUser);
 
 // GET /signup/duplicate_check
 router.get('/signup/duplicate_check', UserController.duplicateUserCheck);
 
-// PUT /preferences
-router.put('/preferences', isAuthenticated, UserController.updatePreferences);
+// POST /verify/send
+router.post('/verify/send', UserController.emailVerificationInitiate);
 
+// GET /verify
+router.get('/verify', UserController.verifyEmail);
+
+/**
+ * ===============
+ * AUTH MANAGEMENT
+ * ===============
+ */
 // POST /reset-password
 router.post('/reset-password', UserController.resetPasswordInitiate);
 
@@ -22,15 +35,20 @@ router.get('/reset-password/:token', UserController.validateResetPasswordToken);
 // POST /reset-password/:token
 router.post('/reset-password/:token', UserController.updatePassword);
 
-// POST /verify/send
-router.post('/verify/send', UserController.emailVerificationInitiate);
-
-// GET /verify
-router.get('/verify', UserController.verifyEmail);
-
 // PUT /account
 router.put('/account', isAuthenticated, UserController.updateSettings);
 
+// DELETE /auth/github
+router.delete('/auth/github', UserController.unlinkGithub);
+
+// DELETE /auth/google
+router.delete('/auth/google', UserController.unlinkGoogle);
+
+/**
+ * ===============
+ * API KEYS
+ * ===============
+ */
 router.post('/account/api-keys', isAuthenticated, UserController.createApiKey);
 
 router.delete(
@@ -39,11 +57,13 @@ router.delete(
   UserController.removeApiKey
 );
 
-// DELETE /auth/github
-router.delete('/auth/github', UserController.unlinkGithub);
-
-// DELETE /auth/google
-router.delete('/auth/google', UserController.unlinkGoogle);
+/**
+ * ===============
+ * USER PREFERENCES
+ * ===============
+ */
+// PUT /preferences
+router.put('/preferences', isAuthenticated, UserController.updatePreferences);
 
 // PUT /cookie-consent
 router.put(
@@ -51,4 +71,5 @@ router.put(
   isAuthenticated,
   UserController.updateCookieConsent
 );
+
 export default router;
